requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.38.37
botocore==1.38.37
Brotli==1.1.0
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
cssselect2==0.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fonttools==4.58.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
